pyproject.toml
[tool.poetry]
name = "advent-of-code-2024"
version = "0.1.0"
description = ""
authors = ["lewis.king <lking1221@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.8"
langchain-anthropic = "^0.3.0"
langchain-openai = "^0.2.9"
langchain-community = "^0.3.8"
langchainhub = "^0.1.21"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
